+++ elaston/tools.py
@@ -4,8 +4,8 @@
 
 import numpy as np
 import string
-from pint import Quantity
-from inspect import getfullargspec
+
+from elaston.unit_decorator import units
 
 
 __author__ = "Sam Waseda"
@@ -215,36 +215,3 @@ def _rotate_tensor(tensor, orientation, inverse, axes=None):
         *len(axes) * [orthonormalize(orientation)],
         v,
     ).reshape(tensor.shape)
-
-
-def _is_plain(inputs, outputs, args, kwargs):
-    if inputs is None and outputs is None:
-        return True
-    if any([isinstance(arg, Quantity) for arg in args]):
-        return False
-    if any([isinstance(val, Quantity) for val in kwargs.values()]):
-        return False
-    return True
-
-
-def _get_input(kwargs, inputs):
-    kwargs_tmp = kwargs.copy()
-    for key, val in kwargs_tmp.items():
-        if isinstance(val, Quantity):
-            if key in inputs:
-                kwargs_tmp[key] = val.to(inputs[key]).magnitude
-            else:
-                kwargs_tmp[key] = val.magnitude
-    return kwargs_tmp
-
-
-def units(outputs=None, inputs=None):
-    def decorator(func):
-        def wrapper(*args, **kwargs):
-            if _is_plain(inputs, outputs, args, kwargs):
-                return func(*args, **kwargs)
-            kwargs.update(zip(getfullargspec(func).args, args))
-            result = func(**_get_input(kwargs, inputs))
-            return result
-        return wrapper
-    return decorator
